AP_NavEKF: Add vertical velocity reset method for in-flight use.
If the vertical velocity estimate has been badly degraded due to IMU errors, then reset of the height can be ineffective, because a bad vertical velocity estimate can cause the baro  measurements to be rejected shortly after the reset, resulting in a recurring height divergence and reset cycle.

This patch adds a method that resets the vertical velocity state to the GPS if available or to the baro measurement otherwise. We try to avoid reset using a baro sourced height rate if GPS vertical velocity is available, because if the height source timeout was caused by a baro spike, then the baro height rate could be corrupted.